Move stageTransaction to mutation
